requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
geojson==3.2.0
idna==3.10
oauthlib==3.3.1
pyowm==3.5.0
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
tweepy==4.16.0
urllib3==2.5.0
